Remove graph display refresh request system
I don't remember what exactly it was supposed to fix, but removing subscription calling Sigma.refresh()
didn't cause any issues with adding, removing or changing colors of the elements. Even before that, graph
refresh wasn't requested in multiple functions changing colors of elements. It is possible that,
whatever the issue that caused implementing it was, it was fixed at some point in development.